+++ cosmohub/api/hadoop/hdfs.py
@@ -2,6 +2,15 @@ import collections
 import copy
 import io
 import os
+import pkg_resources
+import struct
+import thriftpy2
+
+from thriftpy2.transport import TMemoryBuffer
+from thriftpy2.http import TFileObjectTransport
+from thriftpy2.protocol import TCompactProtocol
+
+parquet_thrift = thriftpy2.load(pkg_resources.resource_filename('cosmohub.resources', 'parquet.thrift'), module_name="parquet_thrift")
 
 class HDFSPathReader(io.RawIOBase):
     """\
@@ -163,3 +172,84 @@ class HDFSPathReader(io.RawIOBase):
         `tell()` and `truncate()` will raise IOError.
         """
         return True
+
+class HDFSParquetReader(HDFSPathReader):
+    """\
+    Read an HDFS Parquet dataset (either file or directory) as a stream of bytes.
+  
+    If the path refers to a directory, the contents of the files (ordered by
+    name) are concatenated in the resulting stream. 
+    The offsets of each FileMetaData are adjusted and merged into one.
+    """
+
+    def _initialize(self):
+        """\
+        Open the requested path and compute some internal parameters.
+
+        Those parameters are needed to calculate the final size of the stream or
+        to be able to seek inside the stream.
+        """
+        self._header = b''
+        self._footer = b''
+        
+        self._length = 0
+        self._position = 0
+        self._all_files = collections.deque()
+        
+        fmd_merged = None
+        offset = 0
+
+        status = self._client.status(self._path)
+        if status['type']=='FILE' and status['length']>0:
+            self._all_files.append({
+                'name': os.path.basename(self._path),
+                'length': status['length'],
+                'offset': 0,
+            })
+            self._path = os.path.dirname(self._path)
+            self._length = status['length']
+
+        else:
+            self._header = b'PAR1'
+
+            for _, entry in self._client.list(self._path, status=True):
+                if entry['type'] != 'FILE' or entry['length']==0:
+                    continue
+                
+                with self._client.read(os.path.join(self._path, entry['pathSuffix']), offset=entry['length']-8, length=4) as fd:
+                    fmd_len = struct.unpack('<i', fd.read(4))[0]
+                
+                with self._client.read(os.path.join(self._path, entry['pathSuffix']), offset=entry['length']-fmd_len-8, length=fmd_len) as fd:
+                    tbuf = TFileObjectTransport(fd)
+                    tprot = TCompactProtocol(tbuf)
+                    tfmd = parquet_thrift.FileMetaData()
+                    tfmd.read(tprot)
+                    
+                    if fmd_merged is None:
+                        fmd_merged = copy.deepcopy(tfmd)
+                        offset += entry['length'] - fmd_len -12
+                    else:
+                        for rg in tfmd.row_groups:
+                            rg = copy.deepcopy(rg)
+                            for c in rg.columns:
+                                if c.file_offset:
+                                    c.file_offset += offset
+                                if c.meta_data.data_page_offset:
+                                    c.meta_data.data_page_offset += offset
+                            fmd_merged.row_groups.append(rg)
+                        offset += entry['length'] - fmd_len -12
+                
+                self._all_files.append({
+                    'name': entry['pathSuffix'],
+                    'length': entry['length'] - fmd_len - 8,
+                    'offset': 4,
+                })
+                self._length += entry['length'] - fmd_len - 12
+            
+            tmem = TMemoryBuffer()
+            tprot = TCompactProtocol(tmem)
+            fmd_merged.write(tprot)
+            self._footer = tmem.getvalue()
+            self._footer += struct.pack('<i', len(self._footer)) + b'PAR1'
+        
+        self._current_files = copy.deepcopy(self._all_files)